[WIP] Update demo notebook to run using new masking code
Serialization of masking and weights, and weights implementation of losses is still pending. Notice that implementing weighting could potentially solve device placing and sharding for weights (which are used only when __call__ is invoked), instead of masking (which is used when __init__) is invoked. Nonetheless, sharding information would have to be saved somewhere within model.GraphCast.